requirements.txt
torch
groq
streamlit
flask
flask-jwt-extended
python-telegram-bot
pytest
requests==2.32.3
black
